+++ CHANGELOG.md
@@ -4,8 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [1.0.6 - v1.0.5] - 2026-04-31
-PR#2
+## [1.0.6 - v1.0.5] - 2026-05-02
+PR#3
 
 ### Added
 * `Pipeline/_pipeline.py`
@@ -14,18 +14,35 @@ PR#2
 * Insured Gradient Stability.
 * pipeline Parameter Fix.
 * Model Stability
+* Linear & Logistic Regression Shape varification problem
 
 ----------
 ### New Features
 * `Pipeline Support with Inbuilt Analysis Tool With Regulizations Support`  
+* `Evaluation()` Method support in All `Regression` & `Pipeline`
 
-Read [More About This Parameter](rslearn/Pipeline/README.md) For More Information
+Read [More About This Parameter in Pipeline](rslearn/Pipeline/README.md) For More Information  
+Read [More About This Parameter In linear_models](rslearn/linear_model/README.md)  For More Information  
 
 ### Notes
 ***`See Older Version for more Information`***
 
 ### Changed 
 * License MIT to GNU GPL v3
+* analysis() function's name to `evaluate()`
 
 ### File Changed
-[ALL] due to License
+[ALL] due to License & Fixes
+
+
+### Initial Commit File Changed
+- linear_model/_LinearRegression.py
+- linear_model/_LogisticRegression.py
+- linear_model/_regulizations.py
+- linear_model/README.md
+- Pipeline/_pipeline.py
+- Pipeline/README.md
+- __init__.py
+- CHANGELOG.md
+- .gitignore
+- README.md